We weren't communicating seeds before performing initialisation. This means for runs with 'seed device' or 'seed time' in the input file each processor obtained a different random see.

+++ hydro.h
@@ -308,7 +308,7 @@ typedef struct {
 
   /** Random seed so it is kept the same across nodes.
    */
-  long seed;
+  int seed;
 
   /** What is used to source gravitational waves? (fluid, scalar or
    *  both)
